Separate rendering & logic into their own different steps
Context:: the game loop runs continously, without any sleep or pause.
At each iteration the `index` of the current frame is calculated as: `index = (currentTimeMs - t0TimeMs) / msPerOneFrame`.
The index of the previous frame is also rememebered.

So a check is done to see if the current iteration `lies` inside the previous frame still and if so, everything is skipped.

Previously this was done only inside the `logic` step (`Process()`) , now it is added for the rendering step too;