Major overhaul: 3
Dungeon & server info is partially read from DUNGEON_STATIC_DATA.json & SERVER_STATIC_DATA.json .
(world server still makes a query to the database for some dungeon info, that will be removed so as to rely only on the static files).

Re-implemented skill hotkeys, with the same mechanism as in the old game.

Updated 'Release' configurations for all projects.

Most of the dungeons are opened now. Simple dungeons, passages, villages and even sieges (sieges are treated as 'simple' dungeons for now, i.e. the siege war logic is currently disabled)